Added support for hidden option for documentation
Currently there's no way of avoding the exposure of an entity attribute.
Sometimes there are internal and conditional attributes that are not
important to be known in the documentation to whoever is consuming that
entity.

In this case, we are simply checking if there's a `hidden` attribute in
the `documentation` options and we skip it when parsing the entity
attributes.

Signed-off-by: Vítor Avelino <vavelino@suse.com>