Properly render `example` for array exposure done using another entity (#68)
* Properly render example for arrays of subresources
* Move array typing to data_type_from

This is so it matches waht happens in entity_model_type.

The main difference is we use Symbols (both keys and values) here, whereas we use
Strings there. I'm not sure the difference is intentional, but I didn't want to
introduce any actual changes in this commit.

* DRY up AttributeParser#call

The reason I'm opening this PR is because logic that was present for "standard" exposures
was accidentally left out from "using: Entity..." exposures". The point of this commit
is to hopefully prevent such omissions from being made again.