Allow Proc for enum values in documentation (#77)
* Allow Proc for enum values in documentation

Enables the use of a Proc for the `values` option within
Grape::Entity documentation. This allows for dynamic generation of
enum values based on runtime logic when generating Swagger
specifications.

For example, this is useful when you want to specify a model's enum
for the `values` option and need to lazy-load the model, like this:

```ruby
expose :some, documentation: { values: -> { Model.some_enum.keys } }
```

This provides consistency, as Proc is also allowed for the `values`
option in request parameters:
https://github.com/ruby-grape/grape-swagger/blob/3e842a138b0f5edd2538d9da4aa23276e63e5d5c/lib/grape-swagger/doc_methods/parse_params.rb#L179-L180

* Set enum only if the return value of Proc is an Array