Make entity fields required by default (#81)
This PR changes how grape-swagger-entity determines if an entity field
is required or not:
1. if `documentation: { required: true/false }` is set, use that
2. otherwise, if `if` or `unless` is not null or if `expose_nil` is
`false`, assume the parameter is not required
3. otherwise, assume the parameter is required

This is consistent with how grape-entity renders fields: fields are
always exposed unless `if` or `unless` is provided (cf.
https://github.com/ruby-grape/grape-entity?tab=readme-ov-file#conditional-exposure).

Closes #80.

This is probably a breaking change, so I understand if you don't want to
merge this as is, but it would be nice to be able to opt into this
behavior.